Safari-609.1.11.1/LayoutTests/http/tests/dom/noopener-window-not-targetable2-expected.txt

Make sure that windows opened with 'noopener' via window.open are not targetable. If testing manually, you should see 2 tabs open.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS w is null
PASS w.location.href is "about:blank"
PASS testRunner.windowCount() is 3
PASS successfullyParsed is true

TEST COMPLETE
